% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{ezql_assign_primary_key}
\alias{ezql_assign_primary_key}
\title{Assign Primary Key to a SQL Table if None Exists}
\usage{
ezql_assign_primary_key(
  table,
  primary_key_column,
  schema = NULL,
  database = NULL,
  address = NULL
)
}
\arguments{
\item{table}{A string specifying the name of the SQL table to modify.}

\item{primary_key_column}{A string specifying the name of the column to set as the primary key.}

\item{schema}{A string specifying the schema. Defaults to `NULL` and uses the value from `ezql_details_schema()` if not provided.}

\item{database}{A string specifying the database name. Defaults to `NULL` and uses the value from `ezql_details_db()` if not provided.}

\item{address}{A string specifying the SQL server address. Defaults to `NULL` and uses the value from `ezql_details_add()` if not provided.}
}
\value{
A list containing the result of the primary key assignment query, or a warning if the primary key already exists.
}
\description{
This function assigns a primary key to a specified SQL table if it doesn't already have one.
}
