% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{ezql_change_names}
\alias{ezql_change_names}
\title{Change Data Frame Column Names Using a Rosetta Stone}
\usage{
ezql_change_names(df, rosetta, from = "from", to = "to")
}
\arguments{
\item{df}{A data frame whose column names need to be changed.}

\item{rosetta}{A data frame containing at least two columns: one for the current names and another for the new names. The names of these columns are specified by `from` and `to`.}

\item{from}{A string specifying the name of the column in `rosetta` that contains the current names. Defaults to "from".}

\item{to}{A string specifying the name of the column in `rosetta` that contains the new names. Defaults to "to".}
}
\value{
A data frame with the column names changed according to the mapping provided in `rosetta`.
}
\description{
This function changes the column names of a data frame based on a mapping provided in a "rosetta" table. It uses a custom function to fill missing values in the new column names with the original names if necessary. Optionally, it can drop the column used for filling in the missing names.
}
\examples{
\dontrun{
df <- data.frame(A = 1:3, B = 4:6, C = 7:9)
rosetta <- data.frame(from = c("A", "B"), to = c("X", "Y"))
ezql_change_names(df, rosetta, from = "from", to = "to")
}

}
