% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezql_column.R
\name{ezql_column_sql_add}
\alias{ezql_column_sql_add}
\title{Add a Column to a SQL Table}
\usage{
ezql_column_sql_add(
  table,
  column_name,
  data_type,
  null = TRUE,
  default = NULL,
  backfill_default = FALSE,
  schema = NULL,
  database = NULL,
  address = NULL
)
}
\arguments{
\item{table}{A string specifying the name of the table to modify.}

\item{column_name}{A string specifying the name of the new column to add.}

\item{data_type}{A string specifying the SQL Server data type (e.g., 'VARCHAR(255)', 'INT', 'DATETIME2').}

\item{null}{Logical indicating whether the column should allow NULLs. Default is TRUE.}

\item{default}{A value to set as the default for the column (optional). This will be automatically quoted if it's a character string and not already SQL code (e.g., "hi" becomes "'hi'").}

\item{backfill_default}{Logical. If TRUE, fills existing rows in the table with the default value after adding the column. Default is FALSE.}

\item{schema}{A string specifying the schema. Defaults to NULL and uses the value from ezql_details_schema().}

\item{database}{A string specifying the database name. Defaults to NULL and uses the value from ezql_details_db().}

\item{address}{A string specifying the SQL server address. Defaults to NULL and uses the value from ezql_details_add().}
}
\value{
A list with the result of the query.
}
\description{
Adds a new column to a specified SQL Server table using T-SQL ALTER TABLE.
}
