% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{ezql_connect}
\alias{ezql_connect}
\title{Connect to SQL Server}
\usage{
ezql_connect(database = NULL, address = NULL)
}
\arguments{
\item{database}{The name of the database to connect to. If NULL, it uses the value from `ezql_details_db()`.}

\item{address}{The address of the SQL server. If NULL, it uses the value from `ezql_details_add()`.}
}
\value{
An RODBC connection object if successful; otherwise, an error is thrown.
}
\description{
This function creates a connection to the SQL server using the stored server details.
}
