% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{ezql_delete_data}
\alias{ezql_delete_data}
\title{Delete Rows from a SQL Table}
\usage{
ezql_delete_data(
  df,
  table,
  breakdown = NULL,
  schema = NULL,
  database = NULL,
  address = NULL
)
}
\arguments{
\item{df}{A data frame containing the rows to delete.}

\item{table}{A string specifying the name of the target SQL table.}

\item{breakdown}{A tibble returned by \code{ezql_check_table()} with pre-validated data.
Defaults to \code{NULL}, and validation is performed within the function.}

\item{schema}{A string specifying the schema of the table. Defaults to the value
retrieved from \code{ezql_details_schema()}.}

\item{database}{A string specifying the database name. Defaults to the value
retrieved from \code{ezql_details_db()}.}

\item{address}{A string specifying the server address. Defaults to the value
retrieved from \code{ezql_details_add()}.}
}
\value{
No return value. This function is called for its side effects of deleting
rows from the specified SQL table. If no rows are deleted, a message is printed.
}
\description{
Deletes rows from a SQL table based on a data frame. Ensures that only rows
present in both the data frame and the table are deleted.
}
\details{
If \code{breakdown} is provided, the function assumes that the data frame has
already been validated. Otherwise, it runs \code{ezql_check_table()} to validate
the input.
}
\examples{
\dontrun{
# Delete rows from a table
ezql_delete_data(
  df = rows_to_remove,
  table = "my_table",
  schema = "dbo",
  database = "my_database",
  address = "my_server"
)
}

}
\seealso{
\code{\link{ezql_check_table}}
}
