% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{ezql_details}
\alias{ezql_details}
\title{Get SQL Server Details}
\usage{
ezql_details()
}
\value{
A tibble with the server details.
}
\description{
This function retrieves the stored SQL server details from the session options.
}
