% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{ezql_details_add}
\alias{ezql_details_add}
\title{Get SQL Server Address}
\usage{
ezql_details_add()
}
\value{
A character string representing the server address.
}
\description{
This function retrieves the stored SQL server address from the session options.
}
