% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{ezql_details_db}
\alias{ezql_details_db}
\title{Get SQL Database Name}
\usage{
ezql_details_db()
}
\value{
A character string representing the database name.
}
\description{
This function retrieves the stored SQL database name from the session options.
}
