% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{ezql_details_schema}
\alias{ezql_details_schema}
\title{Get SQL Schema Name}
\usage{
ezql_details_schema()
}
\value{
A character string representing the schema name.
}
\description{
This function retrieves the stored SQL schema name from the session options.
}
