% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{ezql_details_to_rprofile}
\alias{ezql_details_to_rprofile}
\title{Add SQL Server Details to .Rprofile}
\usage{
ezql_details_to_rprofile(schema, database, address)
}
\arguments{
\item{schema}{The schema name.}

\item{database}{The database name.}

\item{address}{The server address.}
}
\description{
This function adds the SQL server details to the .Rprofile file.
}
