% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{ezql_drop}
\alias{ezql_drop}
\title{Drop a SQL Table}
\usage{
ezql_drop(table, schema = NULL, database = NULL, address = NULL)
}
\arguments{
\item{table}{A string specifying the name of the SQL table to be dropped.}

\item{schema}{A string specifying the schema of the table. Defaults to the value
retrieved from \code{ezql_details_schema()}.}

\item{database}{A string specifying the database name. Defaults to the value
retrieved from \code{ezql_details_db()}.}

\item{address}{A string specifying the server address. Defaults to the value
retrieved from \code{ezql_details_add()}.}
}
\value{
Invisible \code{NULL}. The function is called for its side effect of
deleting the specified table from the database.
}
\description{
Deletes a specified table from a SQL database.
}
\details{
This function constructs and executes a SQL \code{DROP TABLE} query to remove
the specified table from the database. If the table does not exist or the query
fails, an error is raised.
}
\examples{
\dontrun{
# Drop a table from the database
ezql_drop(
  table = "my_table",
  schema = "dbo",
  database = "my_database",
  address = "my_server"
)
}

}
\seealso{
\code{\link{ezql_query}}, \code{\link{ezql_details_schema}}, \code{\link{ezql_details_db}}
}
