% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{ezql_full_table_name}
\alias{ezql_full_table_name}
\title{Construct Full Table Name for SQL Queries}
\usage{
ezql_full_table_name(schema, table)
}
\arguments{
\item{schema}{A string specifying the schema name.}

\item{table}{A string specifying the table name.}
}
\value{
A string representing the full table name in the format `schema.table`, which is suitable for use in SQL queries.
}
\description{
This function constructs the full table name, including the schema, for use in SQL queries.
}
\details{
The function combines the schema and table names with a period (.) to create a fully qualified table name that can be used in SQL queries. This is useful for ensuring that the correct table is referenced, especially in databases with multiple schemas.
}
