% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{ezql_list_tables}
\alias{ezql_list_tables}
\title{List Tables on the SQL Server}
\usage{
ezql_list_tables(schema = NULL, database = NULL, address = NULL)
}
\arguments{
\item{schema}{Optional. Schema name. Defaults to `ezql_details_schema()`.}

\item{database}{Optional. Database name. Defaults to `ezql_details_db()`.}

\item{address}{Optional. SQL server address. Defaults to `ezql_details_add()`.}
}
\value{
A tibble with columns: `TABLE_NAME`, `TABLE_TYPE`, and optionally others.
}
\description{
Returns a tibble of table names available in the current schema, database, and address context.
}
