% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{ezql_log_path}
\alias{ezql_log_path}
\title{Retrieve the Configured Logging Path}
\usage{
ezql_log_path(warn = TRUE)
}
\arguments{
\item{warn}{Logical; whether to issue a warning if the logging path is not set.
Defaults to \code{TRUE}.}
}
\value{
A character string containing the logging path, or \code{NULL} if none exists.
}
\description{
This function retrieves the current logging path set in R options.
If no logging path has been defined (e.g., via \code{ezql_log_path_to_rprofile()}),
it returns \code{NULL} and optionally warns the user.
}
\examples{
\dontrun{
ezql_log_path()
}

}
