% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{ezql_log_path_to_rprofile}
\alias{ezql_log_path_to_rprofile}
\title{Add Logging Path to .Rprofile}
\usage{
ezql_log_path_to_rprofile(log_path, overwrite = FALSE)
}
\arguments{
\item{log_path}{A string specifying the full path to the log file.}

\item{overwrite}{Logical; whether to overwrite an existing log path entry in
the .Rprofile file. Defaults to `FALSE`.}
}
\description{
This function writes the logging file path to the user's .Rprofile file,
enabling persistent logging configuration across R sessions.
It will *not* overwrite an existing path unless explicitly allowed.
}
