% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{ezql_new_internal}
\alias{ezql_new_internal}
\title{Internal Function to Create a New Table in the Database}
\usage{
ezql_new_internal(
  df,
  table,
  rosetta,
  rosetta_names_col = "sql_name",
  rosetta_sql_types_col = "sql_type",
  schema = NULL,
  database = NULL,
  address = NULL,
  primary_key = NULL
)
}
\arguments{
\item{df}{A data frame containing the data to be inserted into the new table.}

\item{table}{A string specifying the name of the table to be created.}

\item{rosetta}{A data frame that maps R column names to SQL column names and SQL types. This mapping ensures the correct data types are used in the SQL table.}

\item{rosetta_names_col}{A string specifying the column name in the rosetta data frame that contains the SQL column names. Defaults to `"sql_name"`.}

\item{rosetta_sql_types_col}{A string specifying the column name in the rosetta data frame that contains the SQL data types. Defaults to `"sql_type"`.}

\item{schema}{A string specifying the schema name. Defaults to `NULL` and uses the value from `ezql_details_schema()` if not provided.}

\item{database}{A string specifying the database name. Defaults to `NULL` and uses the value from `ezql_details_db()` if not provided.}

\item{address}{A string specifying the SQL server address. Defaults to `NULL` and uses the value from `ezql_details_add()` if not provided.}

\item{primary_key}{A string specifying the column to set as the primary key. Defaults to `NULL` and no primary key is set if not provided.}
}
\value{
No return value. This function is called for its side effects of creating a table in the specified schema and database.
}
\description{
This internal function creates a new table in the specified schema and database using the provided data frame and a rosetta data frame that maps R column names to SQL column names and types.
}
\details{
This function connects to the specified SQL server, creates a new table according to the schema provided in the rosetta data frame, and inserts the data from the provided data frame. The function will stop and throw an error if there are any issues with the rosetta mapping, such as duplicate column names or mismatches with the data frame.
}
\keyword{internal}
