% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{ezql_primary_key_data}
\alias{ezql_primary_key_data}
\title{Retrieve Primary Key Data from a SQL Table}
\usage{
ezql_primary_key_data(table, schema = NULL, database = NULL, address = NULL)
}
\arguments{
\item{table}{A string specifying the name of the SQL table.}

\item{schema}{A string specifying the schema of the table. Defaults to the value
retrieved from \code{ezql_details_schema()}.}

\item{database}{A string specifying the database name. Defaults to the value
retrieved from \code{ezql_details_db()}.}

\item{address}{A string specifying the server address. Defaults to the value
retrieved from \code{ezql_details_add()}.}
}
\value{
A tibble containing the data for the primary key column(s).
}
\description{
Fetches the data for the primary key column(s) from a specified table in a SQL database.
}
\details{
This function retrieves the primary key column(s) from a specified table
and returns their data as a tibble. If the table does not have a primary key,
the function stops with an error.
}
\examples{
\dontrun{
# Retrieve primary key data
pk_data <- ezql_primary_key_data(
  table = "my_table",
  schema = "dbo",
  database = "my_database",
  address = "my_server"
)
print(pk_data)
}

}
\seealso{
\code{\link{ezql_primary_key_name}}, \code{\link{ezql_query}}
}
