% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{ezql_query}
\alias{ezql_query}
\title{Execute SQL Queries on a Server}
\usage{
ezql_query(queries, database = NULL, address = NULL)
}
\arguments{
\item{queries}{A character vector of SQL queries to execute.}

\item{database}{A string specifying the database name. Defaults to `NULL` and uses `ezql_details_db()` if not provided.}

\item{address}{A string specifying the SQL Server address. Defaults to `NULL` and uses `ezql_details_add()` if not provided.}
}
\value{
A list of results for each query. Each element contains either:
  - `list(success = TRUE, result = <data or success message>)`, or
  - `list(success = FALSE, error = <error message>)`
}
\description{
Executes one or more SQL queries on a specified SQL Server and returns results or errors.
}
