% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{ezql_replace}
\alias{ezql_replace}
\title{Replace an Existing Table in the Database}
\usage{
ezql_replace(
  df,
  table,
  rosetta,
  primary_key = NULL,
  schema = NULL,
  database = NULL,
  address = NULL
)
}
\arguments{
\item{df}{A data frame containing the data to be inserted into the new table.}

\item{table}{A string specifying the name of the table to be replaced.}

\item{rosetta}{A data frame that maps R column names to SQL column names and types.}

\item{primary_key}{A string specifying the column to set as the primary key. Defaults to `NULL` and no primary key is set if not provided.}

\item{schema}{A string specifying the schema name. Defaults to `NULL` and uses the value from `ezql_details_schema()` if not provided.}

\item{database}{A string specifying the database name. Defaults to `NULL` and uses the value from `ezql_details_db()` if not provided.}

\item{address}{A string specifying the SQL server address. Defaults to `NULL` and uses the value from `ezql_details_add()` if not provided.}
}
\value{
No return value. Stops with an error message if the table does not exist.
}
\description{
This function replaces an existing table in the specified schema and database with a new one created from the provided data frame. It stops with an error if the table does not exist.
}
\details{
This function checks for the existence of the specified table, drops it if it exists, and creates a new table using the `ezql_new_internal` function.
}
