% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezql_column.R
\name{ezql_rosetta}
\alias{ezql_rosetta}
\title{Retrieve Ezekiel Rosetta Data}
\usage{
ezql_rosetta()
}
\value{
A tibble containing the data from the "ezql_rosetta" SQL table.
}
\description{
This function retrieves data from the SQL table named "ezql_rosetta".
}
