% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{ezql_set_data_types}
\alias{ezql_set_data_types}
\title{Set Data Types for a DataFrame Based on a Rosetta Mapping}
\usage{
ezql_set_data_types(
  df,
  rosetta,
  names_column = "names",
  type_function_column = "type_function"
)
}
\arguments{
\item{df}{A dataframe whose columns are to be transformed.}

\item{rosetta}{A dataframe containing at least two columns: one with the names of the columns to transform (`names_column`) and another with the names of the functions to apply (`type_function_column`).}

\item{names_column}{A string specifying the name of the column in `rosetta` that contains the column names of `df`. Defaults to "names".}

\item{type_function_column}{A string specifying the name of the column in `rosetta` that contains the function names as strings. Defaults to "type_function".}
}
\value{
A dataframe with the specified columns transformed according to the functions in `rosetta`. The dataframe will have its original grouping restored if it was grouped.
}
\description{
This function applies type transformations to columns in a dataframe based on a provided mapping (rosetta). The function checks for the presence of the specified columns and functions, applies the transformations, and then restores any previous grouping on the dataframe.
}
\examples{
\dontrun{
df <- tibble::tibble(A = c("1", "2", "3"), B = c("TRUE", "FALSE", "TRUE"))
rosetta <- tibble::tibble(names = c("A", "B"), type_function = c("as.numeric", "as.logical"))
df <- ezql_set_data_types(df, rosetta)
}

}
