% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{ezql_set_details}
\alias{ezql_set_details}
\title{Set SQL Server Details}
\usage{
ezql_set_details(schema = NULL, database = NULL, address = NULL)
}
\arguments{
\item{schema}{The schema name (default: NULL).}

\item{database}{The database name (default: NULL).}

\item{address}{The server address (default: NULL).}
}
\description{
This function sets the SQL server details in the session options.
}
