% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{ezql_table}
\alias{ezql_table}
\title{Retrieve and Prepare a SQL Table using Rosetta}
\usage{
ezql_table(
  table,
  as_of = NULL,
  schema = NULL,
  database = NULL,
  address = NULL,
  use_rosetta = TRUE,
  names_column = "sql",
  rename_to = "r",
  type_function_column = "r_type"
)
}
\arguments{
\item{table}{Name of the SQL table to fetch.}

\item{as_of}{Optional. A date to filter the records by `ValidFrom` and `ValidTo`.}

\item{schema}{Optional. Schema name. Defaults to `ezql_details_schema()`.}

\item{database}{Optional. Database name. Defaults to `ezql_details_db()`.}

\item{address}{Optional. SQL server address. Defaults to `ezql_details_add()`.}

\item{use_rosetta}{Logical. Whether to apply column renaming and typing using `ezql_rosetta()`. Defaults to `TRUE`.}

\item{names_column}{Column in rosetta that matches SQL names. Defaults to "sql".}

\item{rename_to}{Column in rosetta to rename columns to. Defaults to "r".}

\item{type_function_column}{Column in rosetta with coercion functions. Defaults to "r_type".}
}
\value{
A tibble with optionally renamed and typed columns.
}
\description{
Fetches a SQL table and optionally filters to records valid as of a given date.
Applies column renaming and type coercion using the `ezql_rosetta()` mapping.
}
