% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{ezql_table_exists}
\alias{ezql_table_exists}
\title{Check if a Table Exists in the Database}
\usage{
ezql_table_exists(table, schema = NULL, database = NULL, address = NULL)
}
\arguments{
\item{table}{A string specifying the name of the table to check for existence.}

\item{schema}{A string specifying the schema name. Defaults to `NULL` and uses the value from `ezql_details_schema()` if not provided.}

\item{database}{A string specifying the database name. Defaults to `NULL` and uses the value from `ezql_details_db()` if not provided.}

\item{address}{A string specifying the SQL server address. Defaults to `NULL` and uses the value from `ezql_details_add()` if not provided.}
}
\value{
Logical `TRUE` if the table exists, `FALSE` otherwise.
}
\description{
This function checks if a specified table exists in the given schema within the database.
}
\details{
This function checks the existence of a table by querying the `INFORMATION_SCHEMA.TABLES` view in the specified schema and database.
}
