% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{ezql_table_names}
\alias{ezql_table_names}
\title{Retrieve Column Names of a SQL Table}
\usage{
ezql_table_names(table, schema = NULL, database = NULL, address = NULL)
}
\arguments{
\item{table}{A string specifying the name of the SQL table.}

\item{schema}{A string specifying the schema of the SQL table. Defaults to the value
retrieved from \code{ezql_details_schema()}.}

\item{database}{A string specifying the database name. Defaults to the value
retrieved from \code{ezql_details_db()}.}

\item{address}{A string specifying the server address. Defaults to the value
retrieved from \code{ezql_details_add()}.}
}
\value{
A character vector of column names in the specified table.
}
\description{
This function retrieves the column names of a specified table from a SQL database.
}
\details{
The function constructs and executes a SQL query to fetch the column names of the specified
table. If \code{schema}, \code{database}, or \code{address} are not explicitly provided,
they are inferred from predefined global defaults.
}
\examples{
\dontrun{
# Retrieve column names from a table
column_names <- ezql_table_names(
  table = "my_table",
  schema = "dbo",
  database = "my_database",
  address = "my_server"
)
print(column_names)
}

}
\seealso{
\code{\link{ezql_query}}, \code{\link{ezql_connect}}
}
