% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{ezql_add_data}
\alias{ezql_add_data}
\title{Add Data to a SQL Table}
\usage{
ezql_add_data(
  df,
  table,
  breakdown = NULL,
  schema = NULL,
  database = NULL,
  address = NULL
)
}
\arguments{
\item{df}{A data frame containing the data to add.}

\item{table}{A string specifying the name of the target SQL table.}

\item{breakdown}{A tibble returned by \code{ezql_check_table()} with pre-validated data.
Defaults to \code{NULL}, and validation is performed within the function.}

\item{schema}{A string specifying the schema of the table. Defaults to the value
retrieved from \code{ezql_details_schema()}.}

\item{database}{A string specifying the database name. Defaults to the value
retrieved from \code{ezql_details_db()}.}

\item{address}{A string specifying the server address. Defaults to the value
retrieved from \code{ezql_details_add()}.}
}
\value{
No return value. This function is called for its side effects of adding
data to the specified SQL table. If no rows are added, a message is printed.
}
\description{
Appends data from a data frame to a specified table in a SQL database.
Ensures that the column names in the data frame match those in the table.
}
\details{
If \code{breakdown} is provided, the function assumes that the data frame has
already been validated. Otherwise, it runs \code{ezql_check_table()} to validate
the input.
}
\examples{
\dontrun{
# Add rows to a table
ezql_add_data(
  df = my_data,
  table = "my_table",
  schema = "dbo",
  database = "my_database",
  address = "my_server"
)
}

}
\seealso{
\code{\link{ezql_check_table}}
}
