% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{ezql_get}
\alias{ezql_get}
\title{Fetch Data from SQL Table}
\usage{
ezql_get(table, schema = NULL, database = NULL, address = NULL)
}
\arguments{
\item{table}{The name of the SQL table.}

\item{schema}{The schema name (default: NULL, will use ezql_details_schema()).}

\item{database}{The database name (default: NULL, will use ezql_details_db()).}

\item{address}{The server address (default: NULL, will use ezql_details_add()).}
}
\value{
A tibble containing the data from the specified SQL table.
}
\description{
This function fetches all data from the specified SQL table.
}
