% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{ezql_make_temporal}
\alias{ezql_make_temporal}
\title{Make a Table Temporal in SQL Server}
\usage{
ezql_make_temporal(
  table,
  primary_key_column,
  schema = NULL,
  database = NULL,
  address = NULL
)
}
\arguments{
\item{table}{A string specifying the name of the SQL table to convert to temporal.}

\item{primary_key_column}{A string specifying the name of the column to set as the primary key. Defaults to "ID".}

\item{schema}{A string specifying the schema. Defaults to `NULL` and uses the value from `ezql_details_schema()` if not provided.}

\item{database}{A string specifying the database name. Defaults to `NULL` and uses the value from `ezql_details_db()` if not provided.}

\item{address}{A string specifying the SQL server address. Defaults to `NULL` and uses the value from `ezql_details_add()` if not provided.}
}
\value{
No return value. The function executes SQL queries to alter the table.
}
\description{
This function converts a specified SQL table into a system-versioned temporal table.
}
