% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{ezql_new}
\alias{ezql_new}
\title{Create a New Table in the Database if it Does Not Exist}
\usage{
ezql_new(
  df,
  table,
  rosetta,
  rosetta_names_col = "sql_name",
  rosetta_sql_types_col = "sql_type",
  primary_key = NULL,
  schema = NULL,
  database = NULL,
  address = NULL
)
}
\arguments{
\item{df}{A data frame containing the data to be inserted into the new table.}

\item{table}{A string specifying the name of the table to be created.}

\item{rosetta}{A data frame that maps R column names to SQL column names and types.}

\item{rosetta_names_col}{A string specifying the column name in the rosetta data frame that contains the SQL column names. Defaults to `"sql_name"`.}

\item{rosetta_sql_types_col}{A string specifying the column name in the rosetta data frame that contains the SQL data types. Defaults to `"sql_type"`.}

\item{primary_key}{A string specifying the column to set as the primary key. Defaults to `NULL` and no primary key is set if not provided.}

\item{schema}{A string specifying the schema name. Defaults to `NULL` and uses the value from `ezql_details_schema()` if not provided.}

\item{database}{A string specifying the database name. Defaults to `NULL` and uses the value from `ezql_details_db()` if not provided.}

\item{address}{A string specifying the SQL server address. Defaults to `NULL` and uses the value from `ezql_details_add()` if not provided.}
}
\value{
No return value. Stops with an error message if the table already exists.
}
\description{
This function creates a new table in the specified schema and database using the provided data frame if the table does not already exist.
}
\details{
This function checks for the existence of the specified table and only creates it if it does not exist, using the `ezql_new_internal` function for the actual creation process.
}
